pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cleartrace"
version = "0.7.0"
description = "ClearTrace — record, replay, explain, and improve LLM agent behavior — deterministic testing + black-box explainability in one library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Joshua", email = "joshuanishanth120@gmail.com" },
]
keywords = [
    "agents",
    "llm",
    "testing",
    "pytest",
    "regression",
    "replay",
    "record",
    "vcr",
    "deterministic",
    "ai",
    "explainability",
    "xai",
    "lime",
    "interpretability",
    "prompt-engineering",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "deepdiff>=7.0",
    "click>=8.0",
    "rich>=13.0",
    "anyio>=4.0",
    "pyyaml>=6.0",
    "wrapt>=1.16",
]

[project.optional-dependencies]
# --- LLM providers ---
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
litellm = ["litellm>=1.0"]
# --- Frameworks ---
langchain = ["langchain-core>=0.3"]
langgraph = ["langgraph>=0.2", "langchain-core>=0.3"]
crewai = ["crewai>=0.30"]
# --- Record add-ons ---
rag = []
ragas = ["ragas>=0.2.0", "langchain-openai>=0.1.0"]
deepeval = ["deepeval>=1.0.0"]
mcp = []
semantic = ["openai>=1.0.0"]
# --- Vector stores ---
chromadb = ["chromadb>=0.5.0"]
pinecone = ["pinecone>=3.0.0"]
qdrant = ["qdrant-client>=1.9.0"]
weaviate = ["weaviate-client>=4.0.0"]
llamaindex = ["llama-index-core>=0.10.0"]
eval = ["openai>=1.0"]
# --- Explain engine ---
explain = ["numpy>=1.24", "pydantic>=2.0", "jinja2>=3.1"]
advisor = ["numpy>=1.24", "pydantic>=2.0", "jinja2>=3.1"]
charts = ["matplotlib>=3.7", "numpy>=1.24", "pydantic>=2.0", "jinja2>=3.1"]
word = ["python-docx>=1.0", "numpy>=1.24", "pydantic>=2.0", "jinja2>=3.1"]
# --- Everything ---
all = [
    "openai>=1.0",
    "anthropic>=0.30",
    "litellm>=1.0",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "crewai>=0.30",
    "ragas>=0.2.0",
    "langchain-openai>=0.1.0",
    "deepeval>=1.0.0",
    "chromadb>=0.5.0",
    "pinecone>=3.0.0",
    "qdrant-client>=1.9.0",
    "weaviate-client>=4.0.0",
    "llama-index-core>=0.10.0",
    "numpy>=1.24",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "matplotlib>=3.7",
    "python-docx>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "openai>=1.0",
    "anthropic>=0.30",
    "numpy>=1.24",
    "pydantic>=2.0",
    "jinja2>=3.1",
]

[project.entry-points."pytest11"]
cleartrace = "clear_trace.pytest_plugin"

[project.scripts]
cleartrace = "clear_trace.cli:main"

[project.urls]
Homepage = "https://github.com/ioteverythin/TraceOps"
Documentation = "https://github.com/ioteverythin/TraceOps/blob/main/ARCHITECTURE.md"
Repository = "https://github.com/ioteverythin/TraceOps"
Issues = "https://github.com/ioteverythin/TraceOps/issues"
Changelog = "https://github.com/ioteverythin/TraceOps/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/clear_trace"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["clear_trace"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "pass",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501", "A002", "N803"]

[tool.ruff.lint.per-file-ignores]
"tests/test_integration_*.py" = ["E402", "B017"]
